# docs/api-reference/v1/openapi-v1.json
{
  "openapi": "3.0.3",
  "servers": [
    {
      "url": "https://api.cal.com/v1"
    }
  ],
...
  "info": {
    "title": "@calcom/api: Public API for Cal.com",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "apiKey"
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/teams": {
      "get": {
        "operationId": "listTeams",
        "summary": "Find all teams",
        "parameters": [
          {
            "in": "query",
            "name": "apiKey",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your API key"
          }
        ],
        "tags": ["Teams"],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization information is missing or invalid."
          },
          "404": {
            "description": "No teams were found"
          }
        }
      },
      "post": {
        "operationId": "addTeam",
        "summary": "Creates a new team",
        "parameters": [
          {
            "in": "query",
            "name": "apiKey",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your API key"
          }
        ],
        "requestBody": {
          "description": "Create a new custom input for an event type",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "slug",
                  "hideBookATeamMember",
                  "brandColor",
                  "darkBrandColor",
                  "timeZone",
                  "weekStart",
                  "isPrivate"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the team"
                  },
                  "slug": {
                    "type": "string",
                    "description": "A unique slug that works as path for the team public page"
                  },
                  "hideBookATeamMember": {
                    "type": "boolean",
                    "description": "Flag to hide or show the book a team member option"
                  },
                  "brandColor": {
                    "type": "string",
                    "description": "Primary brand color for the team"
                  },
                  "darkBrandColor": {
                    "type": "string",
                    "description": "Dark variant of the primary brand color for the team"
                  },
                  "timeZone": {
                    "type": "string",
                    "description": "Time zone of the team"
                  },
                  "weekStart": {
                    "type": "string",
                    "description": "Starting day of the week for the team"
                  },
                  "isPrivate": {
                    "type": "boolean",
                    "description": "Flag indicating if the team is private"
                  },
                  "ownerId": {
                    "type": "number",
                    "description": "ID of the team owner - only admins can set this."
                  },
                  "parentId": {
                    "type": "number",
                    "description": "ID of the parent organization."
                  }
                }
              }
            }
          }
        },
        "tags": ["Teams"],
        "responses": {
          "201": {
            "description": "OK, team created"
          },
          "400": {
            "description": "Bad request. Team body is invalid."
          },
          "401": {
            "description": "Authorization information is missing or invalid."
          }
        }
      }
    },
    "/teams/{teamId}": {
      "delete": {
        "operationId": "removeTeamById",
        "summary": "Remove an existing team",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID of the team to delete"
          },
          {
            "in": "query",
            "name": "apiKey",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your API key"
          }
        ],
        "tags": ["Teams"],
        "responses": {
          "201": {
            "description": "OK, team removed successfully"
          },
          "400": {
            "description": "Bad request. Team id is invalid."
          },
          "401": {
            "description": "Authorization information is missing or invalid."
          }
        }
      },
      "get": {
        "operationId": "getTeamById",
        "summary": "Find a team",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID of the team to get"
          },
          {
            "in": "query",
            "name": "apiKey",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your API key"
          }
        ],
        "tags": ["Teams"],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authorization information is missing or invalid."
          },
          "404": {
            "description": "Team was not found"
          }
        }
      },
      "patch": {
        "operationId": "editTeamById",
        "summary": "Edit an existing team",
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID of the team to edit"
          },
          {
            "in": "query",
            "name": "apiKey",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your API key"
          }
        ],
        "requestBody": {
          "description": "Create a new custom input for an event type",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the team"
                  },
                  "slug": {
                    "type": "string",
                    "description": "A unique slug that works as path for the team public page"
                  }
                }
              }
            }
          }
        },
        "tags": ["Teams"],
        "responses": {
          "201": {
            "description": "OK, team edited successfully"
          },
          "400": {
            "description": "Bad request. Team body is invalid."
          },
          "401": {
            "description": "Authorization information is missing or invalid."
          }
        }
      }
    },
...
  }
}